Rewrite to be path-oriented and much faster
We now make a single Conduit API call to retrieve the full set of
projects for all input paths. This requires us to do path matching
ourselves, which makes this code more complex, but the overall speed
benefit is worth it.

We now also print the results organized by path and grouped by common
ownership. This path-oriented format is easier to scan that the previous
project-oriented output. It also makes it more obvious how many review
groups each file touches. The ownership strength of each project is also
displayed.